Address review: rename methods, merge test files, add TODO
- Rename _get_gateway_provider → _get_mlflow_gateway_provider
- Rename _run_tool_calling_loop → _invoke_and_handle_tools
- Merge test_gateway_invocation.py into test_gateway_adapter.py since
  gateway_invocation.py no longer exists
- Add TODO for extending _call_llm_provider_api with tools support
